refactor(seeds): reorganize insertion to separate context
