+++ api/db/seeds/data/common/constants.js
@@ -1,14 +1,35 @@
 const DEFAULT_PASSWORD = 'pix123';
 const COMMON_OFFSET_ID = 1000;
 
-// FEATURES
+//USERS
+const USER_ID_ADMIN_ORGANIZATION = COMMON_OFFSET_ID;
+const USER_ID_MEMBER_ORGANIZATION = COMMON_OFFSET_ID + 1;
+
+// ADMIN USER ID
+const REAL_PIX_SUPER_ADMIN_ID = 90000;
+
+// TARGET PROFILE
+const PIX_PUBLIC_TARGET_PROFILE_ID = 76;
+
+// FEATURES TABLE
 const FEATURE_COMPUTE_ORGANIZATION_LEARNER_CERTIFICABILITY_ID = COMMON_OFFSET_ID + 1;
 const FEATURE_MULTIPLE_SENDING_ASSESSMENT_ID = COMMON_OFFSET_ID + 2;
 const FEATURE_PLACES_MANAGEMENT_ID = COMMON_OFFSET_ID + 3;
 const FEATURE_MISSIONS_MANAGEMENT_ID = COMMON_OFFSET_ID + 4;
 const FEATURE_LEARNER_IMPORT_ID = COMMON_OFFSET_ID + 5;
 const FEATURE_CAN_REGISTER_FOR_A_COMPLEMENTARY_CERTIFICATION_ALONE_ID = COMMON_OFFSET_ID + 6;
-const ONDE_IMPORT_FORMAT_ID = COMMON_OFFSET_ID + 6;
+
+
+//ORGANIZATIONS
+const SCO_MANAGING_ORGANIZATION_ID = COMMON_OFFSET_ID;
+const SUP_MANAGING_ORGANIZATION_ID = COMMON_OFFSET_ID + 1;
+const SCO_FREGATA_MANAGING_ORGANIZATION_ID = COMMON_OFFSET_ID + 2;
+const SCO_ORGANIZATION_ID = COMMON_OFFSET_ID + 3;
+const SUP_ORGANIZATION_ID = COMMON_OFFSET_ID + 4;
+const PRO_ORGANIZATION_ID = COMMON_OFFSET_ID + 5;
+
+// ORGANIZATION FEATURES TABLE
+const IMPORT_FORMAT_ONDE_ID = COMMON_OFFSET_ID + 1;
 
 //TAGS
 const AGRICULTURE_TAG = {
@@ -48,19 +69,6 @@ const LYCEE_TAG = {
   name: 'LYCEE',
 };
 
-//USERS
-const USER_ID_ADMIN_ORGANIZATION = COMMON_OFFSET_ID;
-const USER_ID_MEMBER_ORGANIZATION = COMMON_OFFSET_ID + 1;
-
-//ORGANIZATIONS
-const SCO_MANAGING_ORGANIZATION_ID = COMMON_OFFSET_ID;
-const SUP_MANAGING_ORGANIZATION_ID = COMMON_OFFSET_ID + 1;
-const SCO_FREGATA_MANAGING_ORGANIZATION_ID = COMMON_OFFSET_ID + 2;
-
-const SCO_ORGANIZATION_ID = COMMON_OFFSET_ID + 3;
-const SUP_ORGANIZATION_ID = COMMON_OFFSET_ID + 4;
-const PRO_ORGANIZATION_ID = COMMON_OFFSET_ID + 5;
-
 export {
   AEFE_TAG,
   AGRICULTURE_TAG,
@@ -73,13 +81,15 @@ export {
   FEATURE_MISSIONS_MANAGEMENT_ID,
   FEATURE_MULTIPLE_SENDING_ASSESSMENT_ID,
   FEATURE_PLACES_MANAGEMENT_ID,
+  IMPORT_FORMAT_ONDE_ID,
   LYCEE_TAG,
   MEDNUM_TAG,
-  ONDE_IMPORT_FORMAT_ID,
+  PIX_PUBLIC_TARGET_PROFILE_ID,
   POLE_EMPLOI_TAG,
   PRIVE_TAG,
   PRO_ORGANIZATION_ID,
   PUBLIC_TAG,
+  REAL_PIX_SUPER_ADMIN_ID,
   SCO_FREGATA_MANAGING_ORGANIZATION_ID,
   SCO_MANAGING_ORGANIZATION_ID,
   SCO_ORGANIZATION_ID,
