+++ api/db/seeds/data/team-acces/build-archived-organizations.js
@@ -1,4 +1,4 @@
-import { REAL_PIX_SUPER_ADMIN_ID } from '../common/common-builder.js';
+import { REAL_PIX_SUPER_ADMIN_ID } from '../common/constants.js';
 
 export function buildArchivedOrganizations(databaseBuilder) {
   databaseBuilder.factory.buildOrganization({
