+++ api/db/seeds/data/team-prescription/data-builder.js
@@ -1,6 +1,5 @@
 import { buildCampaigns } from './build-campaigns.js';
 import { buildOrganizationLearners } from './build-learners.js';
-import { buildOrganizationLearnerImportFormat } from './build-organization-learner-import-formats.js';
 import { buildPlaceLots } from './build-place-lots.js';
 import { buildTargetProfiles } from './build-target-profiles.js';
 
@@ -8,7 +7,6 @@ async function teamPrescriptionDataBuilder({ databaseBuilder }) {
   await buildTargetProfiles(databaseBuilder);
   await buildCampaigns(databaseBuilder);
   await buildOrganizationLearners(databaseBuilder);
-  await buildOrganizationLearnerImportFormat(databaseBuilder);
   await buildPlaceLots(databaseBuilder);
 }
